Move IndexIterator out of NDArray.kt.
